email-parser/requirements.txt
annotated-types
anyio
async_lru
cachetools
certifi
charset-normalizer
click
colorama
fastapi
google-api-core
google-api-python-client
google-cloud-aiplatform
google-auth
google-auth-httplib2
googleapis-common-protos
h11
httplib2
idna
proto-plus
protobuf
pyasn1
pyasn1_modules
pydantic
pydantic-settings
pydantic_core
pyparsing
python-dotenv
requests
rsa
sniffio
slowapi
starlette
tenacity
typing_extensions
uritemplate
urllib3
uvicorn
PyYAML
